refactor scheduler settings
