+++ services/director-v2/src/simcore_service_director_v2/core/settings.py
@@ -26,7 +26,6 @@ from pydantic import (
     ConstrainedStr,
     Field,
     NonNegativeInt,
-    PositiveFloat,
     parse_obj_as,
     validator,
 )
@@ -46,6 +45,7 @@ from settings_library.utils_logging import MixinLoggingSettings
 from simcore_postgres_database.models.clusters import ClusterType
 from simcore_sdk.node_ports_v2 import FileLinkType
 
+from .dynamic_services.scheduler import DynamicServicesSchedulerSettings
 from .dynamic_sidecar_settings import DynamicSidecarSettings
 
 _MINUTE: Final[NonNegativeInt] = 60
@@ -76,22 +76,6 @@ class DirectorV0Settings(BaseCustomSettings):
             path=f"/{self.DIRECTOR_V0_VTAG}",
         )
         return url
-
-
-class DynamicServicesSchedulerSettings(BaseCustomSettings):
-    DIRECTOR_V2_DYNAMIC_SCHEDULER_ENABLED: bool = True
-
-    DIRECTOR_V2_DYNAMIC_SCHEDULER_INTERVAL_SECONDS: PositiveFloat = Field(
-        5.0, description="interval at which the scheduler cycle is repeated"
-    )
-
-    DIRECTOR_V2_DYNAMIC_SCHEDULER_PENDING_VOLUME_REMOVAL_INTERVAL_S: PositiveFloat = (
-        Field(
-            30 * _MINUTE,
-            description="interval at which cleaning of unused dy-sidecar "
-            "docker volume removal services is executed",
-        )
-    )
 
 
 class DynamicServicesSettings(BaseCustomSettings):
