+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_scheduler.py
@@ -38,8 +38,8 @@ from servicelib.background_task import cancel_task
 from servicelib.fastapi.long_running_tasks.client import ProgressCallback
 from servicelib.fastapi.long_running_tasks.server import TaskProgress
 
+from .....core.dynamic_services.scheduler import DynamicServicesSchedulerSettings
 from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
-from .....core.settings import DynamicServicesSchedulerSettings
 from .....models.dynamic_services_scheduler import SchedulerData, ServiceName
 from ...api_client import SidecarsClient, get_sidecars_client
 from ...docker_api import update_scheduler_data_label
