+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_task.py
@@ -16,7 +16,7 @@ from models_library.users import UserID
 from servicelib.fastapi.long_running_tasks.client import ProgressCallback
 from servicelib.fastapi.long_running_tasks.server import TaskProgress
 
-from ....core.settings import DynamicServicesSchedulerSettings
+from ....core.dynamic_services.scheduler import DynamicServicesSchedulerSettings
 from ._abc import SchedulerInternalsInterface, SchedulerPublicInterface
 from ._core._scheduler import Scheduler
